feat: rename the province to province map
