+++ src/index.ts
@@ -1,5 +1,5 @@
 export { default as Map } from './Map';
-export { default as Province } from './Province';
+export { default as ProvinceMap } from './ProvinceMap';
 export { default as ZonalMap } from './ZonalMap';
 export { default as DistrictMap } from './DistrictMap';
 export type { MapProps, MapDataPoint } from './type/type';
